requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.2.1
pydantic==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.0
selene==2.0.0rc9
selenium==4.17.2
